+++ R/get_marginalmeans.R
@@ -27,7 +27,7 @@ get_marginalmeans <- function(model,
                               ci = 0.95,
                               estimate = getOption("modelbased_estimate", "typical"),
                               transform = NULL,
-                              add_iterations = FALSE,
+                              keep_iterations = FALSE,
                               verbose = TRUE,
                               ...) {
   # check if available
@@ -253,7 +253,7 @@ get_marginalmeans <- function(model,
         estimate = estimate,
         datagrid = datagrid,
         transform = !is.null(transform),
-        add_iterations = add_iterations
+        keep_iterations = keep_iterations
       )
     )
   )
@@ -382,8 +382,7 @@ get_marginalmeans <- function(model,
   c(
     "at", "by", "focal_terms", "adjusted_for", "predict", "trend", "comparison",
     "contrast", "estimate", "p_adjust", "transform", "datagrid", "preserve_range",
-    "coef_name", "slope", "ci", "model_info", "contrast_filter", "add_iterations",
-    "posterior_draws"
+    "coef_name", "slope", "ci", "model_info", "contrast_filter", "keep_iterations"
   )
 }