+++ tests/testthat/test-estimate_predicted.R
@@ -113,13 +113,13 @@ test_that("estimate_expectation - Bayesian", {
       chains = 2
     )
   )
-  estim <- estimate_link(model, add_iterations = TRUE)
+  estim <- estimate_link(model, keep_iterations = TRUE)
   draws <- bayestestR::reshape_iterations(estim)
   expect_equal(c(nrow(draws), ncol(draws)), c(2000, 8))
 
   # Non-sampling algorithms
   model <- rstanarm::stan_glm(mpg ~ disp, data = mtcars, algorithm = "meanfield", refresh = 0)
-  estim <- estimate_link(model, add_iterations = TRUE)
+  estim <- estimate_link(model, keep_iterations = TRUE)
   expect_equal(dim(estim), c(10, 1005))
 })