+++ NEWS.md
@@ -9,8 +9,6 @@
   The former `"average"` option is still available, but now returns marginal
   means fully averaged across the sample.
 
-* Argument `keep_iterations` was renamed into `add_iterations`.
-
 ## Changes
 
 * The `transform` argument now also works for `estimate_slopes()` and for
@@ -28,7 +26,7 @@
   to calculate predictions on the link-scale and back-transform them to the
   response scale after aggregation by groups.
 
-* `estimate_means()` gets a `add_iterations` argument, to keep all posterior
+* `estimate_means()` gets a `keep_iterations` argument, to keep all posterior
   draws from Bayesian models as attributes, and if numeric, adds columns of
   posterior draws to the output.